♻️ refactor: 너가소개서 결과 페이지 관련 리팩토링
* mock api로 분리되지 않은 부분 수정
* 다른 페이지 코드들과 다르게 불필요한 뎁쓰가 있던 부분 수정
* 리액트 쿼리 및 에러바운더리 먹여두기
