+++ src/infrastructure/remote/neoga.ts
@@ -1,4 +1,5 @@
 import { NeogaService } from '@api/neoga';
+import { NotFoundError } from '@api/types/errors';
 import { AxiosError } from 'axios';
 import { NEOGA_DATA } from '../mock/neoga.data';
 import { privateAPI } from './base';
@@ -170,6 +171,48 @@ export function NeogaDataRemote(): NeogaService {
     } else throw '서버 통신 실패';
   };
 
+  const getNeososeoInfo = async (formID: number) => {
+    const response = await privateAPI.get({ url: `/form/detail/${formID}` });
+    if (!response.data) throw new NotFoundError('해당 유저와 폼 아이디로 생성된 폼이 없습니다.');
+    return {
+      id: response.data.id,
+      title: response.data.title,
+      subtitle: response.data.subtitle,
+      darkIconImage: response.data.darkIconImage,
+      createdAt: response.data.createdAt,
+      q: response.data.q,
+      keywordList: response.data.keyword.map((keyword: any) => ({
+        id: keyword.id,
+        content: keyword.name,
+        color: keyword.colorcode,
+      })),
+    };
+  };
+
+  const getNeososeoFeedback = async (formID: number) => {
+    const response = await privateAPI.get({ url: `/form/detail/${formID}/answer` });
+    return {
+      answerCount: response.data.answerCount,
+      answerList: response.data.answer
+        ? response.data.answer.map((feedback: any) => ({
+            formID: feedback.formID,
+            id: feedback.id,
+            name: feedback.name,
+            relationship: feedback.relationship,
+            content: feedback.content,
+            isPinned: feedback.isPinned,
+            createdAt: feedback.createdAt,
+            keywordList: feedback.keywords.map((keyword: any) => ({
+              id: keyword.id,
+              content: keyword.name,
+              color: keyword.colorcode,
+              answerId: keyword.answerId,
+            })),
+          }))
+        : [],
+    };
+  };
+
   return {
     getBannerTemplate,
     getMainTemplate,
@@ -181,6 +224,8 @@ export function NeogaDataRemote(): NeogaService {
     postAnswerBookmark,
     postCreateForm,
     getCreateFormInfo,
+    getNeososeoInfo,
+    getNeososeoFeedback,
   };
 }
